+++ src/events.rs
@@ -6,11 +6,11 @@ use super::Alias;
 pub struct Event<'a> {
     pub key: &'a str,
     pub name: &'a str,
-    pub range: Range<u32>,
+    pub range: Range<i64>,
     pub alias: Vec<Alias<'a>>,
-    pub races: Vec<u32>,
-    pub parent: Option<u32>,
-    pub children: Vec<u32>,
+    pub races: Vec<usize>,
+    pub parent: Option<usize>,
+    pub children: Vec<usize>,
 }
 
 impl<'a> Ord for Event<'a> {
